Renamed components and made invoice info responsive
